% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rbed2HGVS.R
\name{bedToCds}
\alias{bedToCds}
\title{annotates a bedfile with Transcript, exon & CDS information}
\usage{
bedToCds(bedfile, outname, outdir, db_path)
}
\arguments{
\item{bedfile}{path to bedfile}

\item{outname}{output file name}

\item{outdir}{output file location}

\item{db}{ensembl db object}
}
\value{
writes report
}
\description{
annotates a bedfile with Transcript, exon & CDS information
}
