% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rbed2HGVS.R
\name{bedToTx}
\alias{bedToTx}
\title{annotate bedfile with transcript information}
\usage{
bedToTx(bedfile, db)
}
\arguments{
\item{bedfile}{GRanges}

\item{db}{ensembl db object}
}
\value{
IRanges object
}
\description{
annotate bedfile with transcript information
}
