% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rbed2HGVS.R
\name{combineEndsBed}
\alias{combineEndsBed}
\title{Combines start & end coordinates}
\usage{
combineEndsBed(bed_a, bed_b, bedfile)
}
\arguments{
\item{bedfile}{}
}
\value{
data.frame
}
\description{
Combines start & end coordinates
}
