% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rbed2HGVS.R
\name{getUsDs}
\alias{getUsDs}
\title{getUsDs}
\usage{
getUsDs(missing_tx, bedfile, cds)
}
\arguments{
\item{missing_tx}{character vector of transcript names}

\item{bedfile}{GRanges object containing a single bedfile}

\item{cds}{}
}
\value{

}
\description{
getUsDs
}
