% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rbed2HGVS.R
\name{Rbed2HGVS}
\alias{Rbed2HGVS}
\title{Rbed2HGVS}
\usage{
Rbed2HGVS(bedfile, db)
}
\arguments{
\item{bedfile}{path to bedfile}

\item{db}{path to .sqlite file}
}
\value{
data.frame
}
\description{
Rbed2HGVS
}
