% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rbed2HGVS.R
\name{txToCds}
\alias{txToCds}
\title{annotate bedfile with CDS information}
\usage{
txToCds(results, db)
}
\arguments{
\item{results}{IRanges}

\item{db}{ebsembl db object}
}
\value{
IRanges object
}
\description{
annotate bedfile with CDS information
}
