% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rbed2HGVS.R
\name{Rbed2HGVS}
\alias{Rbed2HGVS}
\title{appends an additional column of interval in HGVS nomenclature to a bedfile}
\usage{
Rbed2HGVS(bedfile, db, preferred_tx = NA)
}
\arguments{
\item{bedfile}{path to bedfile}

\item{db}{ucsc TxDb object}

\item{preferred_tx}{path tsv file where column1="gene symbol", column2="refseq transcript". File should be headerless /
where there are multiple preferred transcripts from a gene, these should one-row per transcript.}
}
\value{
data.frame object with refseqid, gene, exon and hgvs appended
}
\description{
appends an additional column of interval in HGVS nomenclature to a bedfile
}
