% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rbed2HGVS.R
\name{Rbed2HGVS}
\alias{Rbed2HGVS}
\title{appends an additional column of interval in HGVS nomenclature to a bedfile}
\usage{
Rbed2HGVS(bedfile, db = "./data/ucsc_hg19_ncbiRefSeq.sqlite")
}
\arguments{
\item{bedfile}{path to bedfile}
}
\value{
GRanges object with HGVS appended
}
\description{
appends an additional column of interval in HGVS nomenclature to a bedfile
}
