% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rbed2HGVS.R
\name{Rbed2HGVS}
\alias{Rbed2HGVS}
\title{appends an additional column of interval in HGVS nomenclature to a bedfile}
\usage{
Rbed2HGVS(bedfile, preferred_tx = NA, ncores = 1)
}
\arguments{
\item{bedfile}{granges object with seqnames=(1-22,X,Y,MT)}

\item{preferred_tx}{path tsv file where column1="gene symbol", column2="refseq transcript". File should be headerless /
where there are multiple preferred transcripts from a gene, these should one-row per transcript.}

\item{ncores}{number of cores to use for hgvs calculation (default=1)}
}
\value{
list object with [1] data.frame of original bed appended with HGVS [2] given preferred transcripts not available in db [3] preferred transcripts with different version to db
}
\description{
appends an additional column of interval in HGVS nomenclature to a bedfile
}
